alteracao dos inputs e outputs que podem ser infinitos para double ao inves de int - pesou no tempo de execução

+++ src/tabatu_cpp/Tabua.pxd
@@ -8,11 +8,11 @@ cdef extern from "Tabua.h":
     cdef cppclass Tabua:
         Tabua() except +
         Tabua(vector[double] qx)
-        double qx(int x, int t) const
-        double tpx(int x, int t) const
-        double t_qx(int x, int t) const
-        vector[double] qx(int x, vector[int] t) const
-        vector[double] tpx(int x, vector[int] t) const
-        vector[double] t_qx(int x, vector[int] t) const
-        int tempo_futuro_maximo(int x) const
+        double qx(int x, double t) const
+        double tpx(int x, double t) const
+        double t_qx(int x, double t) const
+        vector[double] qx(int x, vector[double] t) const
+        vector[double] tpx(int x, vector[double] t) const
+        vector[double] t_qx(int x, vector[double] t) const
+        double tempo_futuro_maximo(int x) const
         bool possui_fechamento_plato() const
